Calculate view finder in top level view
